Make accelerated time configurable
### What

Make accelerated time configurable like manual close, only for the standalone network.

### Why

Being able to use accelerated time helps when doing manual testing. It's possible to get today without this change by manually changing the stellar-core.cfg in the image, but it would be more convenient if this was a configuration option like manual close. This is only enabled for when running the standalone network, and when not using captive core, in the same way that running standalone is not currently supported with captive core.