Add a GitHub Actions run to replace TravisCI (#251)
Add a GitHub Actions workflow to replace TravisCI.

Travis CI builds have been failing recently with rate limiting from DockerHub. It appears that TravisCI doesn't cache any docker layers across builds, even of popular images like ubuntu:20.04. This means when each build runs it is downloading the ubuntu image each time and hitting the new rate limits Docker Hub introduced in November 2020. GitHub Actions is more efficient, and caches common images. Our tests on GitHub Actions don't download the ubuntu image each time and don't encounter rate limits because of this.

Also, travis-ci.org is shutting down on June 15 and we'd have to make an effort to migrate to travis-ci.com anyway. If we're going to migrate, we may as well migrate to GitHub Actions which we've been using on new repositories and having a lot of success with.

This change doesn't migrate any other non-TravisCI jobs in this repo. There is a CircleCI job that does image publishing, we might like to move that to GitHub Actions, or to Jenkins, but that is out-of-scope of this change.