Add friendbot (#327)
### What
Add friendbot to the quickstart image that runs only for standalone mode.

Add nginx that fronts horizon and friendbot and directs the `/friendbot` endpoint to friendbot and remaining traffic to horizon.

### Why
I frequently use standalone for testing and not having friendbot is inconvenient. It requires me to have different behavior for testnet vs standalone. When testing with testnet my tests and code must use the friendbot fund functionality. When testing with standalone my tests and code must calculate the root key and fund via it, which is complex to get right since tests must coordinate on sequence numbers and tests must have sufficient delay between fundings.

Friendbot is a rather simple application to run in the docker image with no dependencies beyond horizon so the impact and cost to running it are minimal.

I considered two approaches to fronting traffic into horizon and friendbot. The first was to have horizon redirect its /friendbot endpoint to friendbot running on a different port. This is how friendbot is deployed on SDF's testnet. However, that approach is only possible when the absolute URL of friendbot host is known. When running in docker there is no way to know the absolute URL of the docker host. We could assume the absolute URL is `http://localhost:8001` but this would only work with some docker installations, not all, as it makes assumptions about the hostname and host port. So instead I installed nginx and fronted horizon and friendbot on the same port. This is a little more complicated, but it is more reliable.

This change uses supervisor.d's capability to include additional files into its configuration so that we can define a friendbot supervisor.d config file as an extra file, and it is only included in the configuration in standalone when friendbot is needed. As part of this change I shifted the supervisor.d configs for each service into a separate file too so the apps supervisor.d configs are stored in the same ways.

### Backwards compatibility
I've made an effort to make this change as backwards compatible as possible, however it is not entirely backwards compatible with all uses. For example:

1. Since the image is designed to allow users to mount their own configuration files, if someone mounts their own horizon.env it is possible they will run Horizon on port 8000 which will collide with nginx. This will result in horizon failing to start up because nginx will likely start first (it starts up faster) and horizon will fail to start. This should be very easy to discover, as it won't be silent error, the containers logs will identify horizon failed to start and horizon's logs will identify the port is already in use.

2. Someone might be overriding the version environment variables and building the image themselves with different versions. This PR changes to using build args so that the args can be specified outside all stages of the dockerfiles together. This is probably unlikely, but listing for completeness.

If users of the image are running the latest version blindly, without pinning to a specific digest and managing their upgrades methodically, this could cause a problem. If they're managing their upgrades like any software then this should be discovered outside of a production environment.