Fix failing master build (#329)
Use https:// instead of git:// for git docker builds.

The main trigger for this change is that GitHub stopped supporting git:// yesterday because it is insecure. I should never have used git:// because it is insecure, so this is a good change regardless.

This change has no functional impact on the build.

Note the .git is required on the tail of the URL as Docker/Podman use it on the URL to identify that this is a git repo and not a tar download.