Make the pubnet CI test more realistic (#333)
Add tests for stellar-core catching up and syncing, and for horizon being up separate from starting ingestion. Use those new tests for all networks (pubnet, testnet, standalone), and stop using the horizon ingestion test for pubnet.

The horizon ingestion test is the only network test run on CI and it runs for all network types (pubnet, testnet, standalone). The volume of transactions on pubnet is so high that it takes the CI build way too long to run for the node to catchup. The main reason we test the image is to verify that it is setup correctly. If the testnet build is able to ingest with horizon there's a good chance the pubnet version will be. But it is still important we test the configuration in the pubnet mode to make sure stellar-core can connect to peers and start catching up, and that horizon still runs and is connected to stellar-core. The new lightweight tests serve this purpose.