Give name to final stage in image (#829)
### What
  Name the final stage in the dockerfile.

  ### Why
The image build process builds multiple independent parts at the same
time and outputs the build logs with the name of each stage that is
being built. All the stages except the last one have a name making it
really easy to follow what's happening during the build. The final stage
has no name and so its logs appear namespaced with "stage-18" which is
not helpful. Naming the last stage will mean the logs will be annotated.