Limit cancel-in-progress for CI to pull requests only (#837)
### What
Make cancel-in-progress conditional on pull request events in the CI
workflow concurrency configuration.

  ### Why
This ensures every commit on protected branches like main runs to
completion while still canceling redundant builds on pull requests. On
the main branch builds run for two reasons: pushes, and scheduled runs.
Pushes build most images, and scheduled runs build only nightly images.
@sisuresh noticed that the main scheduled runs
([eg](https://github.com/stellar/quickstart/actions/runs/19745699438))
are causing the push runs to be canceled because they run for the same
commit. It's a good thing that the workflow causes the push and
scheduled builds on the main branch to be serialised so only one runs at
a time. But a scheduled build shouldn't cancel a push build.