Fix builds when the image json is not compact (#861)
### What
Compact the input json making it a single line for passing into the
GITHUB_ENV that can only support single line, unless more complex
scripting is used.

### Why
I recently made a change to the build workflow that removed a line that
ensured the JSON was compact. The bug doesn't affect builds in this
repository because the images.json is always compacted onto a single
line before it even gets to this step. But builds coming from other
repositories are not, because this is the first step they see and where
the compaction needs to happen.

Failing builds can be seen at:
- https://github.com/stellar/stellar-core/pull/5058
-
https://github.com/stellar/stellar-core/actions/runs/20087758923/job/57628911263?pr=5058